Rename dir to entity-search
