+++ context/app/static/js/pages/TestSearch/TestSearch.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { SearchkitClient, SearchkitProvider } from '@searchkit/client';
 
-import Search from 'js/components/test-search/Search';
+import Search from 'js/components/entity-search/Search';
 
 const skClient = new SearchkitClient();
